remove react{16,17}, update tests
