+++ cli/package.json
@@ -108,9 +108,7 @@
     "vue",
     "react",
     "vue2",
-    "react18",
-    "react17",
-    "react16"
+    "react18"
   ],
   "bin": {
     "cypress": "bin/cypress"
@@ -143,16 +141,6 @@
       "import": "./react/dist/cypress-react.esm-bundler.js",
       "require": "./react/dist/cypress-react.cjs.js",
       "types": "./react/dist/index.d.ts"
-    },
-    "./react16": {
-      "import": "./react16/dist/cypress-react.esm-bundler.js",
-      "require": "./react16/dist/cypress-react.cjs.js",
-      "types": "./react16/dist/index.d.ts"
-    },
-    "./react17": {
-      "import": "./react17/dist/cypress-react.esm-bundler.js",
-      "require": "./react17/dist/cypress-react.cjs.js",
-      "types": "./react17/dist/index.d.ts"
     },
     "./react18": {
       "import": "./react18/dist/cypress-react.esm-bundler.js",
