+++ npm/react/src/index.ts
@@ -2,6 +2,4 @@ export * from './createMount'
 
 export * from './mountHook'
 
-export * from './mountLegacy'
-
 export * from './types'
