--- npm/react16/package.json
@@ -1,58 +1,0 @@
-{
-  "name": "@cypress/react16",
-  "version": "0.0.0-development",
-  "description": "Test React components using Cypress",
-  "main": "dist/cypress-react.cjs.js",
-  "scripts": {
-    "build": "rimraf dist && rollup -c rollup.config.js && mv dist/dist/* dist",
-    "postbuild": "node ../../scripts/sync-exported-npm-with-cli.js",
-    "build-prod": "yarn build",
-    "watch": "yarn build --watch --watch.exclude ./dist/**/*"
-  },
-  "devDependencies": {
-    "@cypress/react": "0.0.0-development",
-    "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-node-resolve": "^11.1.1",
-    "@rollup/plugin-typescript": "8.3.3",
-    "@vitejs/plugin-react": "1.3.1",
-    "cypress": "0.0.0-development",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
-    "rollup": "^2.38.5",
-    "typescript": "^4.2.3"
-  },
-  "peerDependencies": {
-    "@types/react": "16.x",
-    "cypress": "*",
-    "react": "16.x",
-    "react-dom": "16.x"
-  },
-  "files": [
-    "dist"
-  ],
-  "types": "dist/index.d.ts",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/cypress-io/cypress.git"
-  },
-  "homepage": "https://github.com/cypress-io/cypress/blob/master/npm/react18/#readme",
-  "bugs": "https://github.com/cypress-io/cypress/issues/new?assignees=&labels=npm%3A%20%40cypress%2Freact18&template=1-bug-report.md&title=",
-  "keywords": [
-    "react",
-    "cypress",
-    "cypress-io",
-    "test",
-    "testing"
-  ],
-  "unpkg": "dist/cypress-react.browser.js",
-  "module": "dist/cypress-react.esm-bundler.js",
-  "peerDependenciesMeta": {
-    "@types/react": {
-      "optional": true
-    }
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
