--- npm/react16/rollup.config.js
@@ -1,3 +1,0 @@
-import * as RollupConfig from '../react/rollup.config'
-
-export default RollupConfig
