--- npm/react16/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "module": "esnext",
-    "lib": [
-      "es2015",
-      "dom"
-    ],
-    "rootDir": "src",
-    "outDir": "dist",
-    "declaration": true,
-    "strict": true,
-    "types": [
-      "cypress"
-    ],
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "resolveJsonModule": false
-  },
-  "include": ["src/**/*.ts"]
-}
