--- npm/react17/.npmignore
@@ -1,3 +1,0 @@
-examples
-src
-cypress
