--- npm/react17/.releaserc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  ...require('../../.releaserc.base'),
-  branches: [
-    { name: 'master', channel: 'latest' },
-  ],
-}
