+++ system-tests/projects/cypress-in-cypress/src/NewComponent.spec.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { mount } from 'cypress/react17'
+import { mount } from 'cypress/react'
 
 describe('NewComponent', () => {
   it('renders the new component', () => {
