+++ system-tests/projects/react17/cypress/support/component.js
@@ -1,4 +1,4 @@
-import { mount, unmount } from 'cypress/react17'
+import { mount, unmount } from 'cypress/react'
 
 Cypress.Commands.add('mount', mount)
 Cypress.Commands.add('unmount', unmount)
